syslog-ng-ctl: refactor split into many files
