+++ syslog-ng-ctl/CMakeLists.txt
@@ -1,6 +1,20 @@
 set(SYSLOG_NG_CTL_SOURCES
     syslog-ng-ctl.c
     control-client.h
+    commands/commands.h
+    commands/commands.c
+    commands/credentials.h
+    commands/credentials.c
+    commands/verbose.h
+    commands/verbose.c
+    commands/ctl-stats.h
+    commands/ctl-stats.c
+    commands/query.h
+    commands/query.c
+    commands/license.h
+    commands/license.c
+    commands/config.h
+    commands/config.c
     control-client.c
 )
