+++ syslog-ng-ctl/commands/ctl-stats.c
@@ -1,0 +1,44 @@
+/*
+ * Copyright (c) 2019 Balabit
+ *
+ * This library is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 2.1 of the License, or (at your option) any later version.
+ *
+ * This library is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public
+ * License along with this library; if not, write to the Free Software
+ * Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
+ *
+ * As an additional exemption you are allowed to compile & link against the
+ * OpenSSL libraries as published by the OpenSSL project. See the file
+ * COPYING for details.
+ *
+ */
+
+#include "ctl-stats.h"
+
+static gboolean stats_options_reset_is_set = FALSE;
+
+GOptionEntry stats_options[] =
+{
+  { "reset", 'r', 0, G_OPTION_ARG_NONE, &stats_options_reset_is_set, "reset counters", NULL },
+  { NULL,    0,   0, G_OPTION_ARG_NONE, NULL,                        NULL,             NULL }
+};
+
+static const gchar *
+_stats_command_builder(void)
+{
+  return stats_options_reset_is_set ? "RESET_STATS" : "STATS";
+}
+
+gint
+slng_stats(int argc, char *argv[], const gchar *mode, GOptionContext *ctx)
+{
+  return dispatch_command(_stats_command_builder());
+}
